+++ slack-gamebot/models/team.rb
@@ -69,4 +69,12 @@ class Team
     team.save!
     team
   end
+
+  def self.purge!
+    # destroy teams inactive for two weeks
+    Team.where(active: false, :updated_at.lte => 2.weeks.ago).each do |team|
+      Mongoid.logger.info "Destroying #{team}, inactive since #{team.updated_at}, over two weeks ago."
+      team.destroy
+    end
+  end
 end