+++ slack-gamebot/service.rb
@@ -7,6 +7,7 @@ module SlackGamebot
       def start!(team)
         fail 'Token already known.' if @services.key?(team.token)
         EM.next_tick do
+          logger.info "Starting team #{team}."
           server = SlackGamebot::Server.new(team: team)
           LOCK.synchronize do
             @services[team.token] = server